Move to provider model
